Move input parsing to test runner
