+++ src/test/replay_game.h
@@ -5,13 +5,7 @@
 
 #include <SDL3/SDL_stdinc.h>
 
-typedef struct ReplayInput {
-    Uint16 p1;
-    Uint16 p2;
-} ReplayInput;
-
 typedef struct ReplayGame {
-    ReplayInput* inputs;
     size_t start_index;
     Uint8 characters[2];
     Uint8 supers[2];
@@ -20,7 +14,6 @@ typedef struct ReplayGame {
 } ReplayGame;
 
 void ReplayGame_Parse(ReplayGame* game);
-void ReplayGame_Destroy(ReplayGame* game);
 
 #endif
